[REMEDY-142] Convert alerts to notifications
The original issue was that when deleting a remediation playbook, the page would redirect and remove the alert almost immediately. This is because the alert was being rendered inside of the component that was being directed away from. This led me to remove the alert and implement dispatching notifications.

To reproduce:
open details of a Remediation
select Delete remediation from the kebab
delete the remediation

The correct alert is displayed, and the page is redirected, but the alert then immediately disappears.

Besure to test all of the locations where I've changed to `dispatchNotifications`.