fix(REMEDY-360): Use correct status icon in review Actions
To test:

Go to remediations and choose a remediations playbook that has systems with actions that have been remediated. Click that playbook -> click the "Systems" tab -> click the "Issues" column links until you find a system with a "Status" that reads, "Remediated". You'll notice that the bug shows a `TimesIcon` instead of a `CheckIcon`. This PR fixes that.