fix(RHINENG-11503): Playbook log doesn't display properly
Playbook log doesn't display properly after inital load.

To repro:
Automation Toolkit > Remediations > A playbook that has run successfully > Activity > The data that the playbook ran successfully > Direct connected
At this point the playbook log will flash on screen and disappear
Click on the hostname

The issue was that playbookRunSystemDetails is a redux variable that remains after initial load. So when the playbook log is loaded again, the log viewer somehow only captures part of the string or none at all.

The best way to handle this issue is to clear the variable in redux on inital load to ensure there are no incorrect states.