fix(RHINENG-15230): Add ouidID to autoreboot switch (#459)
To test, go to any app where the Remediation modal lives and go through the steps until you get to the Remediation review step. There is an "Auto reboot is on/off" switch. Inspecting it should reveal the new ouiaID.

One place to check is Vulnerability. Simply select a CVE and go to the CVE details table. From the details table, add systems and then click the "Remediate" button to open the Remediate modal.

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>
Co-authored-by: Egor Shamardin <33912805+LightOfHeaven1994@users.noreply.github.com>